requirements.txt
Django==3.0.5
django-autoslug==1.9.7
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
flake8==3.7.9
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.9
virtualenv==20.0.18
